+++ service/member/src/main/java/innercircle/member/application/port/in/AuthUseCase.java
@@ -7,4 +7,6 @@ public interface AuthUseCase {
 
     LoginResponse login(LoginRequest request);
 
+
+
 }
