+++ service/member/src/main/java/innercircle/member/application/port/in/MemberUseCase.java
@@ -1,7 +1,5 @@
 package innercircle.member.application.port.in;
 
-import innercircle.member.application.MemberCreateRequest;
-import innercircle.member.application.MemberCreateResponse;
 import innercircle.member.domain.member.Member;
 
 public interface MemberUseCase {
