+++ service/member/src/test/java/innercircle/member/MemberApplicationTests.java
@@ -1,7 +1,7 @@
 package innercircle.member;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import innercircle.member.application.MemberCreateRequest;
+import innercircle.member.infrastructure.adapter.in.web.dto.MemberCreateRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
